Added extra wall and RNG components
> Adicionadas novas paredes que podem ser ajustadas conforme o desejado e, também, um componente de aleatoriedade à elas;
> Aleatoriedade também adicionada ao spawn do PowerUp, por fins de teste. Objetivo é adicionar esse RNG ao spawn do PowerUp quando um brick for quebrado (WIP).

+++ Paranoid/paranoidProject/duballitem.h
@@ -8,7 +8,7 @@
 class DuBallItem : public DuIObjectItem
 {
 public:
-    DuBallItem(int x, int y, int w, int h, int vx, int vy, int score, QColor color = Qt::green, QGraphicsItem *parent = nullptr);
+    DuBallItem(int x, int y, int w, int h, float vx, float vy, int score, QColor color = Qt::green, QGraphicsItem *parent = nullptr);
     void move() override;
     //int getScore() const { return (int)mscore; }
