+++ Paranoid/paranoidProject/duutil.cpp
@@ -1,6 +1,6 @@
 #include "duutil.h"
 
-int DuUtil::abs(int value)
+float DuUtil::abs(float value)
 {
     return value < 0 ? -value : value;
 }
